functional version of create_training_screen with several aspects to improve UI interaction

+++ lib/domain/models/sets.dart
@@ -1,6 +1,6 @@
 class Sets {
   final int reps;
-  final int weight;
+  final double weight;
 
   Sets({required this.reps, required this.weight});
